Move webui submodule one level down, add notice on how to clone ExtJS with git.
By moving the webui project down the source tree, we (Taodyne) can add some
private stuff to the project. For instance, this commit adds a specific message
during ./configure in case Ext JS is not installed. We don't want this message
to be shown to users that are not within Taodyne.