Strings passed to flight recorder must be available on exit.
I think strdup() without free() is acceptable here, since we'll leak a
limited number of strings. Another option would be add a macro to the flight
recorder, like:
    RECORD(ALWAYS, "File name is", REC_DUP(str));
The REC_DUP macro would strdup() the string and store the pointer in a
map for later deletion (when it becomes useless).