OpenGLState::Light(): create light 0 as needed. Fixes #3311.
The default light is #0. So, any Tao function that affects the current
light shall be able to manipulate light #0 even if no light selection has been
done previously.
That is, the following:
    light_position [...]
...should be equivalent to:
    light 0, false
    light_position [...]
With this commit, the method that handles light manipulations in our GL state
manager will take care of creating light #0 if it is references but does not
exist yet.