OSX: Widget::startRefreshTimer(false) should remove Timer event as well as DisplayLink events. Refs #3306.
On OSX, we do not use Timer events directly to refresh the graphical Widget.
Instead, we have a Display Link thread that generates DisplayLink events,
which are further converted to Timer events.
So when we stop the refresh timer (the DL thread), we should drain both the
DisplayLink events _and_ the Timer events from the event queue.