Add primitives: layout and refresh_also. Update Slides module. Fixes #3320.
- The 'layout' primitive associates a name to the current layout. A layout
may have zero, one or more names. Several layouts may share the same name.
Names are stored in a vector of strings: Layout::names.
- The 'refresh_also' primitive is used to request refresh of other layouts
when the current one is to be refreshed. The names of the target layouts
are stored in a vector of string: Layout::deps.
- After the execution of the XL program is done, the refresh dependencies
are scanned by Layout::CheckRefreshDeps(). Its purpose is to copy refresh
information to the target layouts. For instance in the following example,
CheckRefreshDeps() will detect that the innermost layout (B) refreshes on
time and mouse events, so time and mouse event flags are copied to layout
"A".
  locally
      layout "A"
      writeln "A"
  locally
      seconds
      locally
          writeln "B"
          refresh_also "A"
          mouse_x
- The Slides module now uses 'layout' and 'refresh_also' so that a partial
refresh of code that uses numbered items (+ "Text") will also trigger the
initalization of the item counter. This fixes #3320.