TextSpan::{Save,Restore}: Provide empty overrides for Space() and Bounds(). Fixes #3332.
The default implementation of Space() calls Draw(), which we don't want for the
context-saving classes TextSpan::Save and TextSpan::Restore. Indeed,
TextSpan::Save::Draw() allocates memory (to hold the GL state) that is later
deleted by TextSpan::Restore::Draw(). These methods should therefore always be
called in order, and exactly once, during the drawing phase.
This is not the case if they are called during pagination due to 'rejects'
(when all objects do not fit on one line).
And anyways, Save and Restore have no bounds and take up no space, so they just
have to be ignored during pagination.

{Save,Restore}::Bounds() are not called but I have added them because they
should follow the same logic.