Calibrated the eye distance better, refs #2121 @3h
The objective was to have cubes that looked like cubes, spheres that
looked like spheres, and 45 degree angles with the screen that looked
like 45 degrees and not 10 degrees.

Now, 500 pixels in Z is about the same as 500 pixels in X, and is
about the maximum you can reasonably do on an auto-stereoscopic
screen.

Tested with ISE2014, it actually improves the viewing experience.