Make sure we detect changes in body, closes #3528 @30m
The problem was found in DFCG3's Intro, which has a complex, nested set
of frame textures. The outer frame texture is shared (in movie_reel), i.e.
the same 'self', so we find the same frame. But the inner one (in the Body
argument of movie_reel) may differ from run to run of the outer one.

We failed to re-evaluate Body because we were keeping the old layout.

I wonder if a similar problem could not occur on regular layouts.