Added a way to control number of polygons generated, closes ticket #6
The number of steps is between 1 and 'Max'. It starts at 'Min' for size 1,
and increases by 'Incr' each time size doubles. 

Using Min=1, Incr=2 and Max=25 gives good results for a variety of shapes of
all sizes, with scaling up to 10 (beyond that, you may have to tweak it
yourself).

There is a big difference with the old algorithm if you draw the following:

scale 10000
circle 0, 0, 0.01

The old algorithm was still sending a well-defined circle. The new algorithm
sends a square.